+++ cosmos/profiles/databricks/__init__.py
@@ -1,6 +1,6 @@
 """Databricks Airflow connection -> dbt profile mappings"""
 
-from .client import DatabricksOauthProfileMapping
+from .oauth import DatabricksOauthProfileMapping
 from .token import DatabricksTokenProfileMapping
 
 __all__ = ["DatabricksTokenProfileMapping", "DatabricksOauthProfileMapping"]
